% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-geo-to-main.R
\name{add_geo_to_main_df}
\alias{add_geo_to_main_df}
\title{Add geographic names to main dataset}
\usage{
add_geo_to_main_df(main, geo)
}
\arguments{
\item{main}{data.table containing main dataset with utla23cd column}

\item{geo}{data.table containing UTLA reference data}
}
\value{
data.table with added utla23nm column
}
\description{
Merges UTLA names with main dataset using UTLA codes
}
