% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-tranches.R
\name{add_tranches}
\alias{add_tranches}
\title{Add SSMTR tranches to a data table}
\usage{
add_tranches(
  dt,
  utlacd_col = "utla23cd",
  utlanm_col = "utla23nm",
  tranches_file_path
)
}
\arguments{
\item{dt}{data.table with a column containing UTLA codes}

\item{utlacd_col}{Character name of column containing UTLA codes}

\item{tranches_file_path}{Character file path to tranche assigments data}
}
\value{
data.table with added tranche column
}
\description{
Loads tranche information, add it to dt and cleans up
inconsistencies
}
\details{
Each upper-tier local authority was assigned a "tranche" number in
in 2022 which determined the amount and timing of additional funding they'd
receive for drug and alcohol treatment.

This function adds the tranche number to any data table with a utla23cd
column
}
