% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate-summary-stats.R
\name{calculate_summary_stats}
\alias{calculate_summary_stats}
\title{Calculate summary statistics including quartile ranges}
\usage{
calculate_summary_stats(dt, select_groups = c("year"))
}
\arguments{
\item{dt}{data.table containing analysis data}

\item{rate_col}{Character name of rate column (default: "lab_rate")}

\item{by_col}{Character name of grouping column (default: "year")}
}
\value{
data.table with summary statistics including means, quartiles and ranges
}
\description{
Calculates comprehensive summary statistics for rates including quartile ranges
}
