% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-final-df.R
\name{create_final_df}
\alias{create_final_df}
\title{Create final analysis dataset}
\usage{
create_final_df(sir_df, main_df)
}
\arguments{
\item{sir_df}{data.table containing SIR data}

\item{main_df}{data.table containing main treatment data}
}
\value{
data.table containing merged and prepared data for analysis
}
\description{
Prepares and merges main and SIR data for analysis
}
\details{
\enumerate{
\item Adds geographic names to main data
\item Creates other_ost indicator
\item Merges SIR and main data
}
}
