% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate-subint-count-by-yr.R
\name{estimate_subint_count_by_yr}
\alias{estimate_subint_count_by_yr}
\title{Estimate annual count of clients receiving a subintervention}
\usage{
estimate_subint_count_by_yr(dt, subintervention = "phbudi_any", groupby = NULL)
}
\arguments{
\item{dt}{\code{data.table} with columns client_random, n_jy, submoddt +
a subintervention binary variable and groups required.}

\item{subintervention}{Character the quoted name of the subintervention
column (default: "phbudi_any")}

\item{groupby}{Character the quoted names of desired grouping variables
(default: NULL)}
}
\value{
a data table with 4 (+ length of \code{groupby}) columns
}
\description{
Returns a data table with an estimated count for subintervention
by year
}
\details{
We cannot count by the \code{submoddt} month, since we would be
measuring the number of SIRs per month, not the number of clients.
This count is an \emph{estimate} since it relies on a choice to count a
client with one (or more) SIR indicating the subintervention as
being counted as receiving that subintervention. It is possible,
although unlikely, that a client could be counted after only
receiving the subintervention for a few days.
}
