% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-data-dictionary.R
\name{get_data_dictionary}
\alias{get_data_dictionary}
\title{Get data dictionary}
\usage{
get_data_dictionary(dictionary_xlsx = "data/RAND-data-dictionary.xlsx")
}
\arguments{
\item{dictionary_xlsx}{Path to Excel file containing data dictionary}
}
\value{
data.table containing combined data dictionary
}
\description{
Loads and combines data dictionaries from RAND Excel file
}
\details{
Checks for existing CSV version first
Combines dictionaries from main, SIR, and TOP sheets
Cleans column names
}
