% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-geo-data.R
\name{get_geo_data}
\alias{get_geo_data}
\title{Combine and de-duplicate the 23 and 24 UTLA data}
\usage{
get_geo_data()
}
\value{
data.table containing UTLA codes and names for England
}
\description{
Retrieves UTLA names and codes from official ArcGIS REST API
}
\details{
Filters for English UTLAs (codes starting with 'E')
}
