% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-utla23-data.R
\name{get_utla23_data}
\alias{get_utla23_data}
\title{Get UTLA 2023 geographic reference data}
\usage{
get_utla23_data()
}
\value{
data.table containing UTLA codes and names for England
}
\description{
Retrieves UTLA names and codes from official ArcGIS REST API
}
\details{
Filters for English UTLAs (codes starting with 'E')
}
