% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-main-data.R
\name{load_main_data}
\alias{load_main_data}
\title{Load main treatment data}
\usage{
load_main_data(file_path = "inst/extdata/K3anon_FullDataset_for_VfM.csv")
}
\value{
data.table containing filtered client IDs, journey numbers and UTLA codes
}
\description{
Loads and filters the main treatment dataset for opiate users
}
\details{
Filters for drug_grp == "Opiate" and selects only needed columns
}
