% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-lab-count.R
\name{plot_lab_count}
\alias{plot_lab_count}
\title{Plot LAB treatment counts}
\usage{
plot_lab_count(dt, fill_var = NULL)
}
\arguments{
\item{dt}{data.table containing count data}

\item{fill_var}{Optional character name of variable to use for fill colors}
}
\value{
ggplot object
}
\description{
Creates a column plot of LAB treatment counts
}
\details{
Uses afcharts styling
}
