% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-lab-rate.R
\name{plot_lab_rate}
\alias{plot_lab_rate}
\title{Plot LAB treatment rates}
\usage{
plot_lab_rate(dt, fill_var = NULL)
}
\arguments{
\item{dt}{data.table containing rate data}

\item{fill_var}{Optional character name of variable to use for fill colors}
}
\value{
ggplot object
}
\description{
Creates a column plot of LAB treatment rates as percentages
}
\details{
Uses afcharts styling and formats y-axis as percentages
}
