% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render-data-dictionary.R
\name{render_data_dictionary}
\alias{render_data_dictionary}
\title{Render formatted data dictionary}
\usage{
render_data_dictionary()
}
\value{
gt_table object containing formatted data dictionary
}
\description{
Creates a formatted GT table of the data dictionary
}
\details{
Filters dictionary to only show relevant columns
Formats column names in sentence case
}
