% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unzip-datafile.R
\name{unzip_datafile}
\alias{unzip_datafile}
\title{Unzip inst/extdata files and check contents}
\usage{
unzip_datafile(zipfile = "RAND-LAB-dataset.zip")
}
\arguments{
\item{zipfile}{Character string specifying the name of the zip file (default: "RAND-LAB-inst/extdataset.zip")}
}
\value{
Character vector of CSV filenames in the inst/extdata directory
}
\description{
Unzips the RAND LAB inst/extdataset if not already unzipped and returns list of available files
}
