% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate-aggregation.R
\name{validate_aggregation}
\alias{validate_aggregation}
\title{Validate data aggregation}
\usage{
validate_aggregation(final_df, aggregated_df)
}
\arguments{
\item{final_df}{Pre-aggregation data.table}

\item{aggregated_df}{Post-aggregation data.table}
}
\value{
Invisible NULL, throws assertion error if validation fails
}
\description{
Ensures totals are correct through aggregation steps
}
\details{
This validates the method for taking an annual count of LAB
}
