% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate-rate.R
\name{validate_rate}
\alias{validate_rate}
\title{Validate rate calculations}
\usage{
validate_rate(df)
}
\arguments{
\item{df}{data.table containing rate calculations}
}
\value{
Invisible NULL, throws assertion error if validation fails
}
\description{
Checks validity of calculated rates
}
\details{
Verifies rates are between 0 and 1
Cross-checks rate calculations against raw counts
}
